services/cad/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [ "setuptools>=68", "wheel" ]

[project]
name = "kitty-cad"
version = "0.1.0"
description = "CAD orchestration service for KITTY"
requires-python = ">=3.11"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
]
dependencies = [
  "anyio>=4.4",
  "fastapi>=0.111",
  "httpx>=0.27",
  "minio>=7.2",
  "numpy>=1.26",
  "pydantic>=2.7",
  "Pillow>=10.4",
  "trimesh>=4.0",
  "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
local = [
  "cadquery>=2.5",  # ARM64 wheels not available - use Zoo/Tripo APIs instead
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = [ "src" ]
